requirements.txt
Pillow==10.2.0
reportlab==4.1.0